+++ pycep/typing.py
@@ -31,7 +31,7 @@ ResourceFunctions: TypeAlias = (
     "ExtensionResourceId | ListKeys | PickZones | Reference | ResourceId | SubscriptionResourceId | TenantResourceId"
 )
 ScopeFunctions: TypeAlias = "ManagementGroup | ResourceGroup | Subscription | Tenant"
-StringFunctions: TypeAlias = "Base64 | Base64ToJson | Base64ToString | EndsWith | Format | Guid | IndexOf | LastIndexOf | NewGuid | Split | StartsWith | String | Substring | ToLower | ToUpper | Trim | UniqueString | Uri | UriComponent | UriComponentToString"
+StringFunctions: TypeAlias = "Base64 | Base64ToJson | Base64ToString | DataUri | EndsWith | Format | Guid | IndexOf | LastIndexOf | NewGuid | Split | StartsWith | String | Substring | ToLower | ToUpper | Trim | UniqueString | Uri | UriComponent | UriComponentToString"
 Functions: TypeAlias = "AnyFunctions | ArrayFunctions | DateFunctions | DeploymentFunctions | LogicalFunctions | NumericFunctions | ObjectFunctions | ResourceFunctions | ScopeFunctions | StringFunctions"
 
 
@@ -698,6 +698,15 @@ class Base64ToJson(TypedDict):
 class Base64ToString(TypedDict):
     type: Literal["base64_to_string"]
     parameters: _Base64ToJsonParameters
+
+
+class _DataUriParameters(TypedDict):
+    string_to_convert: PossibleValue
+
+
+class DataUri(TypedDict):
+    type: Literal["data_uri"]
+    parameters: _DataUriParameters
 
 
 class EndsWith(TypedDict):